services/workers/audio_separator_service/requirements.txt
# Audio Separator Service Requirements
# 基于 UVR-MDX 模型的人声/背景音分离服务
# 使用 PaddlePaddle 基础镜像 (CUDA 11.8 + Python 3.10)

# ========================================
# 核心框架依赖
# ========================================
celery>=5.3.0
redis>=5.0.0
pydantic>=2.0.0
PyYAML>=6.0

# ========================================
# 音频分离核心依赖
# ========================================
# audio-separator[gpu] 会自动安装：
# - torch (自动适配 CUDA 11.8)
# - torchaudio
# - onnxruntime-gpu
# 不需要显式声明 PyTorch，避免版本冲突
audio-separator[gpu]

# ========================================
# Demucs 音频分离模型
# ========================================
demucs>=4.0.0

# ========================================
# 音频处理依赖
# ========================================
ffmpeg-python>=0.2.0
librosa>=0.10.0
soundfile>=0.12.0

# ========================================
# 工具库
# ========================================
# 注意: NumPy 已在 PaddlePaddle 基础镜像中预装 (2.2.6)
# audio-separator 会根据需要安装兼容版本
scipy>=1.11.0

# 系统监控库
psutil>=5.9.0
